Refactor: mock user 로직 Handler에서 분리
