+++ webapp/src/mocks/handlers.js
@@ -1,42 +1,10 @@
 import { rest } from 'msw';
 import { ROOT_URL } from 'constant/api';
-import userComments from './user/userComments';
 import teamComments from './team/teamComments';
-import usersList from './user/usersList';
-import userDetail from './user/userDetail';
 import teamDetail from './team/teamDetail';
 import teamsList from './team/teamsList';
 import editTeamDetail from './team/editTeamDetail';
-
-const emptyUsers = {
-  status: '200',
-  isSuccess: true,
-  code: 1000,
-  message: '요청 성공',
-  data: [],
-};
-
-const USER = [
-  rest.get(`${ROOT_URL}/users`, (req, res, ctx) => {
-    const lastPage = Number(req.url.searchParams.get('lastPage'));
-    if (lastPage >= 2) {
-      return res(ctx.status(200), ctx.json(emptyUsers));
-    }
-    return res(ctx.status(200), ctx.json(usersList));
-  }),
-  rest.get(`${ROOT_URL}/user/liking`, (req, res, ctx) => {
-    return res(ctx.status(200), ctx.json(usersList));
-  }),
-  rest.get(`${ROOT_URL}/user/read`, (req, res, ctx) => {
-    return res(ctx.status(200), ctx.json(usersList));
-  }),
-  rest.get(`${ROOT_URL}/user/:id`, (req, res, ctx) => {
-    return res(ctx.status(200), ctx.json(userDetail));
-  }),
-  rest.get(`${ROOT_URL}/user/comment/:id`, (req, res, ctx) => {
-    return res(ctx.status(200), ctx.json(userComments));
-  }),
-];
+import { USER } from './user';
 
 const TEAM = [
   rest.get(`${ROOT_URL}/teams`, (req, res, ctx) => {
